third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/padding-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'padding-top' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
[FAIL] Can set 'padding-left' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
[FAIL] Can set 'padding-right' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
[FAIL] Can set 'padding-bottom' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
Harness: the test ran to completion.
